replace Tabler Icons repo
Ryan doesn't seem to update his repo anymore. Some issues and pull requests to update to the newest icons are stale.
This one is more up to date. It has a weekly automatic GitHub action to update icons from the base repo.
I think this is better and more up to date choice.